Inc. Default speed, simplify speed control
Default drive train speed is now faster, (150 relative unit), and speed control
is not by command anymore, but rather a dv+ or dv- key binding (, or .) which is
the lower shift version of (<,>) on a typical computer keyboard. (So that you
don't have to hold shift when manually debugging the controller over serial). A
relative speed of 150 seems to be a good but upper range of necessary speed,
more testing needed in that direction